refactor: expose permute & fft method in fftinput; update test
