+++ math/src/fft/serial.rs
@@ -7,10 +7,6 @@ use super::fft_inputs::FftInputs;
 use crate::{field::StarkField, utils::log2, FieldElement};
 use utils::{collections::Vec, uninit_vector};
 
-// CONSTANTS
-// ================================================================================================
-const MAX_LOOP: usize = 256;
-
 // POLYNOMIAL EVALUATION
 // ================================================================================================
 
@@ -21,8 +17,8 @@ where
     B: StarkField,
     E: FieldElement<BaseField = B>,
 {
-    fft_in_place(p, twiddles, 1, 1, 0);
-    permute(p);
+    FftInputs::fft_in_place(p, twiddles);
+    FftInputs::permute(p);
 }
 
 /// Evaluates polynomial `p` over the domain of length `p.len()` * `blowup_factor` shifted by
@@ -53,10 +49,10 @@ where
                 *d = (*c).mul_base(factor);
                 factor *= offset;
             }
-            fft_in_place(chunk, twiddles, 1, 1, 0);
+            FftInputs::fft_in_place(chunk, twiddles);
         });
 
-    permute(result.as_mut_slice());
+    FftInputs::permute(result.as_mut_slice());
     result
 }
 
@@ -70,12 +66,10 @@ where
     B: StarkField,
     E: FieldElement<BaseField = B>,
 {
-    fft_in_place(evaluations, inv_twiddles, 1, 1, 0);
     let inv_length = B::inv((evaluations.len() as u64).into());
-
-    // Use fftinputs shift_by on evaluations.
+    FftInputs::fft_in_place(evaluations, inv_twiddles);
     FftInputs::shift_by(evaluations, inv_length);
-    permute(evaluations);
+    FftInputs::permute(evaluations);
 }
 
 /// Interpolates `evaluations` over a domain of length `evaluations.len()` and shifted by
@@ -89,76 +83,11 @@ pub fn interpolate_poly_with_offset<B, E>(
     B: StarkField,
     E: FieldElement<BaseField = B>,
 {
-    fft_in_place(evaluations, inv_twiddles, 1, 1, 0);
-    permute(evaluations);
+    FftInputs::fft_in_place(evaluations, inv_twiddles);
+    FftInputs::permute(evaluations);
 
     let domain_offset = B::inv(domain_offset);
     let offset = B::inv((evaluations.len() as u64).into());
 
-    // Use fftinputs's shift_by_series on evaluations.
     FftInputs::shift_by_series(evaluations, offset, domain_offset);
-}
-
-// PERMUTATIONS
-// ================================================================================================
-
-pub fn permute<B, I>(values: &mut I)
-where
-    B: StarkField,
-    I: FftInputs<B> + ?Sized,
-{
-    let n = values.len();
-    for i in 0..n {
-        let j = super::permute_index(n, i);
-        if j > i {
-            values.swap(i, j);
-        }
-    }
-}
-
-// CORE FFT ALGORITHM
-// ================================================================================================
-
-/// In-place recursive FFT with permuted output.
-///
-/// Adapted from: https://github.com/0xProject/OpenZKP/tree/master/algebra/primefield/src/fft
-pub(super) fn fft_in_place<B, I>(
-    values: &mut I,
-    twiddles: &[B],
-    count: usize,
-    stride: usize,
-    offset: usize,
-) where
-    B: StarkField,
-    I: FftInputs<B> + ?Sized,
-{
-    let size = values.len() / stride;
-    debug_assert!(size.is_power_of_two());
-    debug_assert!(offset < stride);
-    debug_assert_eq!(values.len() % size, 0);
-
-    // Keep recursing until size is 2
-    if size > 2 {
-        if stride == count && count < MAX_LOOP {
-            fft_in_place(values, twiddles, 2 * count, 2 * stride, offset);
-        } else {
-            fft_in_place(values, twiddles, count, 2 * stride, offset);
-            fft_in_place(values, twiddles, count, 2 * stride, offset + stride);
-        }
-    }
-
-    for offset in offset..(offset + count) {
-        I::butterfly(values, offset, stride);
-    }
-
-    let last_offset = offset + size * stride;
-    for (i, offset) in (offset..last_offset)
-        .step_by(2 * stride)
-        .enumerate()
-        .skip(1)
-    {
-        for j in offset..(offset + count) {
-            I::butterfly_twiddle(values, twiddles[i], j, stride);
-        }
-    }
 }
